requirements.txt
py==1.4.31
pytest==3.0.3
python-dateutil==2.5.3
six==1.10.0
wheel==0.24.0
